+++ differences_evaluator/evaluator.py
@@ -1,19 +1,12 @@
 """Differences evaluator."""
-from differences_evaluator import file_parser
-
-
-def print_diff(
-    first_file_path: str,
-    second_file_path: str,
-    format_output_file: str = None
-):
-    print(stylish(generate_diff(first_file_path, second_file_path)))
+from differences_evaluator import file_parser, formaters
 
 
 def generate_diff(
     first_file_path: str,
-    second_file_path: str
-) -> dict:
+    second_file_path: str,
+    format_output_file: str = None
+) -> str:
     """Generate diffirences of two files.
 
     Parameters:
@@ -21,18 +14,27 @@ def generate_diff(
         second_file_path: path to second target file
         format_file: format needed plain or json
     """
+    diff = get_diff(first_file_path, second_file_path)
+
+    if format_output_file is None or format_output_file == 'stylish':
+        return(formaters.stylish(diff))
+    elif format_output_file == 'plain':
+        return formaters.plain(diff)
+
+
+def get_diff(
+    first_file_path: str,
+    second_file_path: str
+) -> dict:
+
     first_file_data = file_parser.get_repr(
         file_parser.get_data(first_file_path)
     )
     second_file_data = file_parser.get_repr(
         file_parser.get_data(second_file_path)
     )
-    if not first_file_data or not second_file_data:
-        return
 
-    diff = get_diff(first_file_data, second_file_data)
-
-    return diff
+    return get_node_diff(first_file_data, second_file_data)
 
 
 def _get_children_force(data: dict, node: str, diff_status: str):
@@ -140,7 +142,7 @@ def _get_children(
         end_data_no_change_node[child['node']] = child
 
     for key in start_data_no_change_node:
-        children.append(get_diff(
+        children.append(get_node_diff(
             node=key,
             start_data=start_data_no_change_node[key],
             end_data=end_data_no_change_node[key]))
@@ -148,7 +150,7 @@ def _get_children(
     return children
 
 
-def get_diff(
+def get_node_diff(
     start_data: dict = None,
     end_data: dict = None,
     node: str = 'root'
@@ -163,6 +165,21 @@ def get_diff(
         formated data with key "diff": add/remove/no change
 
     """
+    def _get_children_force(data: dict, node: str, diff_status: str):
+        result = {'node': node, 'leafs': [], 'children': [], 'diff': diff_status}
+        for key, value in data['leafs'].items():
+            result['leafs'].append(
+                {'key': key, 'value': value, 'diff': diff_status}
+                )
+        if data['children']:
+            for child in data['children']:
+                result['children'].append(_get_children_force(
+                    data=child,
+                    node=child['node'],
+                    diff_status=diff_status
+                ))
+        return result
+
     diff = {'node': node, 'leafs': [], 'children': [], 'diff': 'no change'}
     if start_data['leafs'] or end_data['leafs']:
         diff['leafs'] = _get_leafs_diff(start_data['leafs'], end_data['leafs'])
@@ -176,72 +193,3 @@ def get_diff(
         )
 
     return diff
-
-
-def stylish(diff: dict) -> str:
-    """Format diff data dict.
-     - deleted key: value
-     + added key: value
-       unchanged key: value
-
-    Parameters:
-        diff: differences data representation
-    Returns:
-        formated string
-    """
-
-    def _get_sign(diff_item):
-        if diff_item['diff'] == 'added':
-            return '+'
-        elif diff_item['diff'] == 'removed':
-            return '-'
-        elif diff_item['diff'] == 'no change':
-            return ' '
-
-    def _get_output_stylish_format(value):
-        if value is False:
-            return 'false'
-        elif value is True:
-            return 'true'
-        elif value is None:
-            return 'null'
-        else:
-            return value
-
-    def _get_stylish_node_rows(node, offset=3, force_sign=None):
-        node_rows = []
-        spaces = ' ' * offset
-        for leaf in node['leafs']:
-            sign = _get_sign(leaf)
-            node_rows.append(
-                '{spaces} {sign} {key}: {value}'.format(
-                    spaces=spaces,
-                    sign=force_sign or _get_sign(leaf),
-                    key=leaf['key'],
-                    value=_get_output_stylish_format(leaf['value'])
-                )
-            )
-
-        if node['children']:
-            for child in node['children']:
-                sign = force_sign or _get_sign(child)
-                node_rows.append(
-                    '{spaces} {sign} {node}: '.format(
-                        sign=sign,
-                        spaces=spaces,
-                        node=child['node']
-                    ) + '{'
-                )
-                if force_sign or child['diff'] != 'no change':
-                    node_rows.extend(
-                        _get_stylish_node_rows(child, offset+3, ' ')
-                    )
-                else:
-                    node_rows.extend(_get_stylish_node_rows(child, offset+3))
-        node_rows.append(spaces + '}')
-        return node_rows
-
-    output = ['{']
-    output.extend(_get_stylish_node_rows(diff))
-    output[-1] = '}'
-    return '\n'.join(output)
