+++ differences_evaluator/scripts/gendiff.py
@@ -16,7 +16,7 @@ args = parser.parse_args()
 
 
 def main():
-    evaluator.print_diff(args.first_file, args.second_file, args.format)
+    evaluator.generate_diff(args.first_file, args.second_file, args.format)
 
 
 if __name__ == "__main__":
