+++ tests/fixtures/formaters_content.json
@@ -1,0 +1,62 @@
+{
+    "test_stylish": 
+    {
+        "args": 
+        [
+            {"node": "root",
+                "leafs":[
+                    {"key": "group1", "value": 888, "diff": "added"}],
+                "children": [{"node": "common",
+                    "leafs": [
+                    {"key": "setting1","value": "Value 1","diff": "no change"},
+                    {"key": "setting2", "value": 200, "diff": "no change"},
+                    {"key": "setting3", "value": true, "diff": "removed"},
+                    {"key": "setting3", "value": false, "diff": "added"}],
+                    "children": [
+                        {"node": "setting6", "leafs": [
+                            {"key": "key", "value": "value", "diff": "no change"}],
+                        "children": [
+                        {"node": "doge", "leafs": [
+                            {"key": "wow", "value": "", "diff": "removed"},
+                            {"key": "wow1", "value": "", "diff": "added"}
+                            ],
+                            "children": [], "diff": "no change"}],
+                        "diff": "no change"}],
+                    "diff": "no change"},
+                    {"node": "group1",
+                        "leafs": [{"key": "test", "value": "test", "diff": "removed"}],
+                        "children": [],
+                        "diff": "removed"}
+                    ],
+                "diff": "no change"}
+        ],
+        "expect":
+            [
+                "+ group1: 888",
+                "common: {",
+                "setting1: Value 1",
+                "setting2: 200",
+                "- setting3: true",
+                "+ setting3: false",
+                "setting6: {",
+                "key: value",
+                "doge: {",
+                "- wow:",
+                "+ wow1:",
+                "}",
+                "}",
+                "}",
+                "- group1: {",
+                "test: test",
+                "}"
+            ]
+    },
+    "test_stylish_empty_diff":
+    {
+        "args":
+        [
+            {"node": "root", "leafs": [], "children": [], "diff": "no change"}
+        ],
+        "expect": "{\n}"
+    }
+}
